Update licensing
I initially listed the ISC license for the source code, but it makes no sense for the dataset (which isn't exactly *source code*), and I publicized it so horribly, few would've made the inference it even had a license attached.

This fixes that and moves it to a more sensible licensing scheme. Also, Apache 2.0 is a bit clearer in practice than ISC/MIT for that kind of licensing scheme (and provides a few more protections for everyone all around).